Dev_GUI

+++ src/utils/ConstantsForDrawing.cpp
@@ -6,10 +6,10 @@ namespace DrawConst {
     const float INTERSECTION_POINT_SIZE = 10.f;
     const float POINT_SIZE = 5.f;
 
-    const int WINDOW_WIDTH = 1600;
-    const int WINDOWS_HEIGHT = 900;
-    const int DISPLAY_SIZE = 1600;
-    const int DISPLAY_SCALE = 2;
+    const float WINDOW_WIDTH = 1600;
+    const float WINDOWS_HEIGHT = 900;
+    const float DISPLAY_SIZE = 1600;
+    const float DISPLAY_SCALE = 2;
 
     const float SQUARE_SIDE_SIZE = 800;
     const float INDENT_SIZE = 0;
